Feat:refatoração das funcoes e dashboard  implementada

+++ src/pages/Home/index.jsx
@@ -1,36 +1,56 @@
 import { HomeContainer, GoToLogin } from "../../styles/container";
 import logo from "../../assets/logo.png";
 import { useNavigate } from "react-router-dom";
+import { useContext, useEffect } from "react";
+import { AuthContext } from "../../contexts/AuthContext";
+import { ModalWindow } from "../../components/Modal";
+import { TechList } from "../../components/TechList";
+import { ModalLink } from "../../styles/techlist";
 
 export function Home() {
-  const navigate = useNavigate();
-  const userInfo = {
-    token: localStorage.getItem("KenzieHub:Token"),
-    name: localStorage.getItem("KenzieHub:Name"),
-    module: localStorage.getItem("KenzieHub:Module"),
-  };
-  if (!userInfo.token) navigate("/");
+  // const navigate = useNavigate();
+  // const profile = {
+  //   token: localStorage.getItem("KenzieHub:Token"),
+  //   name: localStorage.getItem("KenzieHub:Name"),
+  //   module: localStorage.getItem("KenzieHub:Module"),
+  // };
+  // if (!profile.token) navigate("/");
 
+  const navigate = useNavigate();
+  const { loading, profile, modal, setModal } = useContext(AuthContext);
+  const userToken = localStorage.getItem("KenzieHub:Token");
+
+  useEffect(() => {
+    if (!userToken) {
+      navigate("/");
+    }
+  }, [navigate, userToken]);
+
+  if (loading) return <h3 className="loading">Loading</h3>;
   return (
-    <HomeContainer>
-      <nav>
-        <img src={logo} alt="Kenzie Hub Logo" />
-        <GoToLogin to={"/"} onClick={() => localStorage.clear()}>
-          Sair
-        </GoToLogin>
-      </nav>
-      <header>
-        <div>
-          <h3>Olá, {userInfo.name}</h3>
-          <p>{userInfo.module}</p>
-        </div>
-      </header>
-      <main>
-        <h3>Que Pena! Estamos em desenvolvimento :(</h3>
-        <p>
-          Nossa aplicação está em desenvolvimento, em breve teremos novidades.
-        </p>
-      </main>
-    </HomeContainer>
+    <>
+      <HomeContainer>
+        {modal ? <ModalWindow /> : null}
+        <nav>
+          <img src={logo} alt="Kenzie Hub Logo" />
+          <GoToLogin to={"/"} onClick={() => localStorage.clear()}>
+            Sair
+          </GoToLogin>
+        </nav>
+        <header>
+          <div>
+            <h3>Olá, {profile.name}</h3>
+            <p>{profile.module}</p>
+          </div>
+        </header>
+        <main>
+          <div className="ulHeader">
+            <h3>Tecnologias</h3>
+            <ModalLink onClick={() => setModal(true)}>+</ModalLink>
+          </div>
+          <TechList techData={profile.techs} />
+        </main>
+      </HomeContainer>
+    </>
   );
 }
